+++ src/plone/testing/tests.py
@@ -1,14 +1,13 @@
 # -*- coding: utf-8 -*-
-try:
-    import unittest2 as unittest
-except ImportError:  # Python 2.7
-    import unittest
 from OFS.SimpleItem import SimpleItem
 from pkg_resources import get_distribution
 from ZPublisher.Iterators import filestream_iterator
+from zope.testing import renormalizing
 
 import doctest
 import os.path
+import re
+import unittest
 import zope.component.testing
 
 
@@ -62,6 +61,22 @@ def tearDown(self):
     zope.component.testing.tearDown()
 
 
+checker = renormalizing.RENormalizing([
+    # normalize py2 output to py3
+    (re.compile(r'__builtin__'), r'builtins'),
+    (re.compile(
+        r"'Unknown directive', u'http://namespaces.zope.org/zope', u'"),
+     r"'Unknown directive', 'http://namespaces.zope.org/zope', '"),
+
+    # normalize py3 output to py3
+    (re.compile(
+        r"zope\.configuration\.xmlconfig\.ZopeXMLConfigurationError"),
+     r"ZopeXMLConfigurationError"),
+    (re.compile(r"builtins\.PopulatedZODB"), r"PopulatedZODB"),
+    (re.compile(r"builtins\.ExpandedZODB"), r"ExpandedZODB"),
+])
+
+
 def test_suite():
     suite = unittest.TestSuite()
     suite.addTests([
@@ -72,6 +87,7 @@ def test_suite():
             'publisher.rst',
             'zodb.rst',
             'z2.rst',
+            checker=checker,
             setUp=setUp,
             tearDown=tearDown,
             optionflags=doctest.ELLIPSIS | doctest.NORMALIZE_WHITESPACE,